requirements.txt
Django>=4.2,<5
psycopg[binary]>=3.1
django-environ>=0.11
requests>=2.31
urllib3>=2.0
discord.py>=2.3.0
python-dateutil>=2.8.0
celery[redis]>=5.3
redis>=5.0

# cppa_slack_transcript_tracker app (Slack Bolt, token extraction, etc.)
slack-bolt>=1.18
pytz>=2024.1
selenium>=4.35

# wg21_paper_tracker app
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-cloud-run>=0.10.1
google-cloud-storage>=2.14.0
